+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - :sparkles: Add `html-validate` module for HTML validation.
 - :rocket: Add a table of contents to the result page.
+- :rocket: Display total number of issues in CLI.
 
 ## [1.4.2] - 2021-01-14